👷 build(commitlint): add mandatory emoji prefix to types
commitlint has hard-coded the prefix "conventional-changelog-" for parserPreset. A wrapper
was added to "conventional-changelog-conventionalcommits", called "conventional-changelog-organon"
to override its header pattern regex.

Currently a link is hard-coded in package.json, but an actual package will have to be produced and
published for the production code to work. Anyway, it's also the case for "@organon/commitlint-config".