👷 build(gulpfile): basic gulpfile to be imported
wip: works fine on self but cannot be linked safely as a gulp instance
is created for each path associated with a require("gulp").
This means that gulp MUST be declared as a peer dependency, and
that, since gulp is also a dev dependency and a gulp package sits within
the node_modules directory, a copy devoid of the node_modules
directory must be linked instead of this dev version.
so doing in the target directory the following cannot work:
npm i --save-dev gulp @organon/gulpfile@link:../gulpfile
Somehow a way must be found for both package and target to depend
on the same gulp package (hoisting? workspace? publishing/installing?)