♻️ refactor(organon-node-rig)!: simplify config
BREAKING CHANGE: no-test profile is removed in favor of
a single tsconfig-no-test.json in the default profile.
config/typescript.json allows to output a different module type
concurrently, so tsconfig-cjs.json is removed.
Tests are now transpiled in lib and cjs outDirs, so they must be
filtered out on publish.