👷 build(rush): check file changes
Based on https://dev.to/kkazala/series/17133:
'Rush and Conventional Commits Series' Articles

As of now, command 'rush changefiles' only goes through git log and
informs on the change status of the files (some parts of the script are
commented out, preventing any edit).

A tweak had to be made to strip emojis from commit types, as
parse-commit-message has hard-coded the shape of a commit type.